alterei o double pelo float

+++ src/Cartao.java
@@ -1,24 +1,24 @@
 public class Cartao {
     private boolean aproximar;
-    private double limite;
-    private double saldo;
+    private float limite;
+    private float saldo;
     public Cartao(){
         this.limite = 2000;
         this.saldo = 1000;
     }
 
-    public double getSaldo() {
+    public float getSaldo() {
         return saldo;
     }
 
-    public void setSaldo(double saldo) {
+    public void setSaldo(float saldo) {
         this.saldo = saldo;
     }
 
-    public Double getLimite() {
+    public float getLimite() {
         return limite;
     }
-    public void setLimite(Double limite) {
+    public void setLimite(float limite) {
 
         this.limite = limite ;
     }
